# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/tasks/{taskId}/dependencies": {
            "get": {
                "description": "Позволяет получить список всех зависимостей задачи. Доступно всем пользователям.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Зависимости задач"
                ],
                "summary": "Получить все зависимости задачи",
                "operationId": "get-dependencies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор задачи",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы",
                        "name": "PageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "По возрастанию/по убыванию",
                        "name": "Order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Характеристика для сортировки",
                        "name": "OrderBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pagination.Paging-models_Dependency"
                        }
                    },
                    "400": {
                        "description": "Некорректный идентификатор",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Задача с заданным идентификатором не найдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Позволяет добавить зависимость у задачи. Доступно всем пользователям, но пользователи могут создавать зависимости только между своими задачами.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Зависимости задач"
                ],
                "summary": "Добавить зависимость у задачи",
                "operationId": "create-dependency",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор задачи",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Dependency"
                        }
                    },
                    "400": {
                        "description": "Некорректный идентификатор",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Задача с заданным идентификатором не найдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tasks/{taskId}/dependencies/{dependencyId}": {
            "delete": {
                "description": "Позволяет удалить зависимость у задачи по идентификатору. Доступно всем пользователям.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Зависимости задач"
                ],
                "summary": "Удалить зависимость",
                "operationId": "delete-dependency",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор задачи",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Идентификатор зависимости",
                        "name": "dependencyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Некорректный идентификатор",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Задача или зависимость с заданным идентификатором не найдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Dependency": {
            "type": "object",
            "required": [
                "child_task_id",
                "parent_task_id"
            ],
            "properties": {
                "child_task_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "parent_task_id": {
                    "type": "integer"
                }
            }
        },
...
        "pagination.Order": {
            "type": "string",
            "enum": [
                "asc",
                "desc"
            ],
            "x-enum-varnames": [
                "Asc",
                "Desc"
            ]
        },
        "pagination.Pagination": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/pagination.PagingOptions"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "pagination.Paging-models_Dependency": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Dependency"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/pagination.Pagination"
                }
            }
        },
...
        "pagination.PagingOptions": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/pagination.Order"
                },
                "order_by": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        }
    }
}